fix(generic-items): resolve PR blockers
