+++ packages/plugin-generic-items/src/generic-items.detail-page.label-entries.ts
@@ -1,0 +1,167 @@
+const translationEntries = <Key extends string>(entries: ReadonlyArray<readonly [Key, string]>) => entries;
+
+export const genericItemsDetailLabelEntries = translationEntries([
+  ['identityTitle', 'cards.basis.identity.title'],
+  ['identityDescription', 'cards.basis.identity.description'],
+  ['metaTitle', 'cards.basis.meta.title'],
+  ['metaDescription', 'cards.basis.meta.description'],
+  ['textTitle', 'cards.content.text.title'],
+  ['textDescription', 'cards.content.text.description'],
+  ['classificationTitle', 'cards.content.classification.title'],
+  ['classificationDescription', 'cards.content.classification.description'],
+  ['relationsTitle', 'cards.content.relations.title'],
+  ['relationsDescription', 'cards.content.relations.description'],
+  ['linksMediaTitle', 'cards.content.linksMedia.title'],
+  ['linksMediaDescription', 'cards.content.linksMedia.description'],
+  ['scheduleTitle', 'cards.content.schedule.title'],
+  ['scheduleDescription', 'cards.content.schedule.description'],
+  ['payloadTitle', 'cards.settings.payload.title'],
+  ['payloadDescription', 'cards.settings.payload.description'],
+  ['secondaryTitle', 'cards.settings.secondary.title'],
+  ['secondaryDescription', 'cards.settings.secondary.description'],
+  ['title', 'fields.title'],
+  ['genericType', 'fields.genericType'],
+  ['visible', 'fields.visible'],
+  ['author', 'fields.author'],
+  ['keywords', 'fields.keywords'],
+  ['externalId', 'fields.externalId'],
+  ['publicationDate', 'fields.publicationDate'],
+  ['publishedAt', 'fields.publishedAt'],
+  ['teaser', 'fields.teaser'],
+  ['categoryName', 'fields.categoryName'],
+  ['categories', 'fields.categories'],
+  ['categoriesHelp', 'fields.categoriesHelp'],
+  ['categoriesSearch', 'fields.categoriesSearch'],
+  ['categoriesSearchPlaceholder', 'fields.categoriesSearchPlaceholder'],
+  ['contacts', 'fields.contacts'],
+  ['webUrls', 'fields.webUrls'],
+  ['addresses', 'fields.addresses'],
+  ['addressAddition', 'fields.addressAddition'],
+  ['addressKind', 'fields.addressKind'],
+  ['street', 'fields.street'],
+  ['zip', 'fields.zip'],
+  ['city', 'fields.city'],
+  ['latitude', 'fields.latitude'],
+  ['longitude', 'fields.longitude'],
+  ['contentBlocks', 'fields.contentBlocks'],
+  ['intro', 'fields.intro'],
+  ['body', 'fields.body'],
+  ['openingHours', 'fields.openingHours'],
+  ['dateFrom', 'fields.dateFrom'],
+  ['dateTo', 'fields.dateTo'],
+  ['timeFrom', 'fields.timeFrom'],
+  ['timeTo', 'fields.timeTo'],
+  ['description', 'fields.description'],
+  ['open', 'fields.open'],
+  ['notAvailable', 'values.notAvailable'],
+  ['weekdayMO', 'values.weekdays.MO'],
+  ['weekdayTU', 'values.weekdays.TU'],
+  ['weekdayWE', 'values.weekdays.WE'],
+  ['weekdayTH', 'values.weekdays.TH'],
+  ['weekdayFR', 'values.weekdays.FR'],
+  ['weekdaySA', 'values.weekdays.SA'],
+  ['weekdaySU', 'values.weekdays.SU'],
+  ['mediaContents', 'fields.mediaContents'],
+  ['mediaCaption', 'fields.mediaCaption'],
+  ['mediaCopyright', 'fields.mediaCopyright'],
+  ['mediaContentType', 'fields.mediaContentType'],
+  ['imageSearch', 'fields.imageSearch'],
+  ['mediaTypeUnspecified', 'values.mediaContentTypes.unspecified'],
+  ['mediaTypeimage', 'values.mediaContentTypes.image'],
+  ['mediaTypeaudio', 'values.mediaContentTypes.audio'],
+  ['mediaTypevideo', 'values.mediaContentTypes.video'],
+  ['mediaTypelogo', 'values.mediaContentTypes.logo'],
+  ['mediaTypeattachment', 'values.mediaContentTypes.attachment'],
+  ['locations', 'fields.locations'],
+  ['locationName', 'fields.locationName'],
+  ['department', 'fields.department'],
+  ['district', 'fields.district'],
+  ['regionName', 'fields.regionName'],
+  ['state', 'fields.state'],
+  ['dates', 'fields.dates'],
+  ['accessibilityInformations', 'fields.accessibilityInformations'],
+  ['priceInformations', 'fields.priceInformations'],
+  ['payload', 'fields.payload'],
+  ['accessibilityTypes', 'fields.accessibilityTypes'],
+  ['accessibilityLinks', 'fields.accessibilityLinks'],
+  ['priceName', 'fields.priceName'],
+  ['priceAmount', 'fields.priceAmount'],
+  ['priceCategory', 'fields.priceCategory'],
+  ['priceDescription', 'fields.priceDescription'],
+  ['groupPrice', 'fields.groupPrice'],
+  ['ageFrom', 'fields.ageFrom'],
+  ['ageTo', 'fields.ageTo'],
+  ['minAdultCount', 'fields.minAdultCount'],
+  ['maxAdultCount', 'fields.maxAdultCount'],
+  ['minChildrenCount', 'fields.minChildrenCount'],
+  ['maxChildrenCount', 'fields.maxChildrenCount'],
+  ['addContentBlock', 'actions.addContentBlock'],
+  ['addCategory', 'actions.addCategory'],
+  ['addAddress', 'actions.addAddress'],
+  ['addOpeningHour', 'actions.addOpeningHour'],
+  ['addImage', 'actions.addImage'],
+  ['addLocation', 'actions.addLocation'],
+  ['addAccessibilityInformation', 'actions.addAccessibilityInformation'],
+  ['addPriceInformation', 'actions.addPriceInformation'],
+  ['uploadMedia', 'actions.uploadMedia'],
+  ['uploadingMedia', 'actions.uploadingMedia'],
+  ['addMediaManual', 'actions.addMediaManual'],
+  ['selectImage', 'actions.selectImage'],
+  ['contentBlockItem', 'content.blockItem'],
+  ['addressItem', 'content.addressItem'],
+  ['openingHourItem', 'content.openingHourItem'],
+  ['locationItem', 'content.locationItem'],
+  ['accessibilityInformationItem', 'content.accessibilityInformationItem'],
+  ['priceInformationItem', 'content.priceInformationItem'],
+  ['removeImage', 'actions.removeImage'],
+  ['removeCategory', 'actions.removeCategory'],
+  ['mediaLibraryDescription', 'cards.content.linksMedia.description'],
+  ['imagePickerEmpty', 'messages.imagePickerEmpty'],
+  ['categoryOptionsLoading', 'messages.categoryOptionsLoading'],
+  ['validationWebUrls', 'validation.webUrls'],
+  ['richTextBlockType', 'richText.blockType'],
+  ['richTextParagraph', 'richText.paragraph'],
+  ['richTextHeading2', 'richText.heading2'],
+  ['richTextHeading3', 'richText.heading3'],
+  ['richTextBlockquote', 'richText.blockquote'],
+  ['richTextBulletList', 'richText.bulletList'],
+  ['richTextOrderedList', 'richText.orderedList'],
+  ['richTextBold', 'richText.bold'],
+  ['richTextItalic', 'richText.italic'],
+  ['richTextUndo', 'richText.undo'],
+  ['richTextRedo', 'richText.redo'],
+  ['richTextApplyLink', 'richText.applyLink'],
+  ['richTextLinkInput', 'richText.linkInput'],
+  ['fields.addressAddition', 'fields.addressAddition'],
+  ['fields.street', 'fields.street'],
+  ['fields.zip', 'fields.zip'],
+  ['fields.city', 'fields.city'],
+  ['fields.latitude', 'fields.latitude'],
+  ['fields.longitude', 'fields.longitude'],
+  ['fields.locationName', 'fields.locationName'],
+  ['fields.department', 'fields.department'],
+  ['fields.district', 'fields.district'],
+  ['fields.regionName', 'fields.regionName'],
+  ['fields.state', 'fields.state'],
+  ['actions.geocodeAddress', 'actions.geocodeAddress'],
+  ['actions.geocodingAddress', 'actions.geocodingAddress'],
+  ['actions.reverseGeocodeAddress', 'actions.reverseGeocodeAddress'],
+  ['actions.reverseGeocodingAddress', 'actions.reverseGeocodingAddress'],
+  ['messages.locationGeocodeDisabled', 'messages.locationGeocodeDisabled'],
+  ['messages.locationGeocodeEmpty', 'messages.locationGeocodeEmpty'],
+  ['messages.locationGeocodeRateLimited', 'messages.locationGeocodeRateLimited'],
+  ['messages.locationGeocodeTimeout', 'messages.locationGeocodeTimeout'],
+  ['messages.locationGeocodeForbidden', 'messages.locationGeocodeForbidden'],
+  ['messages.locationGeocodeUnauthorized', 'messages.locationGeocodeUnauthorized'],
+  ['messages.locationGeocodeError', 'messages.locationGeocodeError'],
+  ['messages.locationMapUnavailable', 'messages.locationMapUnavailable'],
+  ['messages.locationMapError', 'messages.locationMapError'],
+  ['messages.mediaUploadInitializing', 'messages.mediaUploadInitializing'],
+  ['messages.mediaUploadUploading', 'messages.mediaUploadUploading'],
+  ['messages.mediaUploadFinalizing', 'messages.mediaUploadFinalizing'],
+  ['messages.mediaUploadSuccess', 'messages.mediaUploadSuccess'],
+  ['messages.mediaUploadError', 'messages.mediaUploadError'],
+  ['messages.mediaUploadUnsupportedType', 'messages.mediaUploadUnsupportedType'],
+  ['messages.mediaUploadUnavailableUrl', 'messages.mediaUploadUnavailableUrl'],
+  ['validation.geoLocation', 'validation.geoLocation'],
+] as const);
