+++ packages/plugin-generic-items/src/generic-items.detail-page.labels.ts
@@ -1,0 +1,5 @@
+import type { usePluginTranslation } from '@sva/plugin-sdk';
+import { genericItemsDetailLabelEntries } from './generic-items.detail-page.label-entries.js';
+
+export const createGenericItemsDetailLabels = (pt: ReturnType<typeof usePluginTranslation>) =>
+  Object.fromEntries(genericItemsDetailLabelEntries.map(([key, translationKey]) => [key, pt(translationKey)]));
