+++ packages/plugin-generic-items/src/generic-items.geo-fields.shared.tsx
@@ -1,0 +1,2 @@
+export { useGenericItemsGeoFieldState } from './generic-items.geo-field-state.js';
+export { GenericItemsGeoMapSection } from './generic-items.geo-map-section.js';
