+++ packages/plugin-generic-items/src/index.ts
@@ -1,13 +1,6 @@
 export {
-  createGenericItem,
-  deleteGenericItem,
-  getGenericItem,
   GenericItemsApiError,
   listGenericItems,
-  updateGenericItem,
 } from './generic-items.api.js';
-export { GENERIC_ITEMS_CONTENT_TYPE } from './generic-items.constants.js';
-export { GenericItemsDetailPage } from './generic-items.detail-page.js';
 export { GenericItemsCreatePage, GenericItemsEditPage, GenericItemsListPage } from './generic-items.pages.js';
 export { pluginGenericItems, pluginGenericItemsActionDefinitions, pluginGenericItemsPermissionDefinitions } from './plugin.js';
-export type * from './generic-items.types.js';
