+++ packages/plugin-generic-items/src/plugin.translations.de.actions.ts
@@ -1,0 +1,33 @@
+export const genericItemsTranslationsDeActions = {
+  actions: {
+    create: 'Generic Item anlegen',
+    update: 'Änderungen speichern',
+    save: 'Speichern',
+    back: 'Zurück zur Liste',
+    edit: 'Bearbeiten',
+    delete: 'Löschen',
+    deleteConfirm: 'Soll dieses Generic Item wirklich gelöscht werden?',
+    addContentBlock: 'Block hinzufügen',
+    addCategory: 'Kategorie hinzufügen',
+    addAddress: 'Adresse hinzufügen',
+    addOpeningHour: 'Öffnungszeit hinzufügen',
+    addLink: 'Link hinzufügen',
+    addContact: 'Kontakt hinzufügen',
+    addDate: 'Termin hinzufügen',
+    geocodeAddress: 'Geo-Koordinaten ermitteln',
+    geocodingAddress: 'Geo-Koordinaten werden ermittelt',
+    reverseGeocodeAddress: 'Adresse ermitteln',
+    reverseGeocodingAddress: 'Adresse wird ermittelt',
+    addImage: 'Aus Medienbibliothek wählen',
+    addLocation: 'Ort hinzufügen',
+    addAccessibilityInformation: 'Barrierefreiheit hinzufügen',
+    addPriceInformation: 'Preis hinzufügen',
+    uploadMedia: 'Bild hochladen',
+    uploadingMedia: 'Bild wird hochgeladen',
+    addMediaManual: 'Medium manuell ergänzen',
+    selectImage: 'Bild auswählen',
+    removeImage: 'Bild entfernen',
+    removeCategory: 'Kategorie {{name}} entfernen',
+    remove: 'Entfernen',
+  },
+} as const;
