+++ packages/plugin-generic-items/src/plugin.translations.de.messages.ts
@@ -1,0 +1,87 @@
+export const genericItemsTranslationsDeMessages = {
+  content: {
+    blockItem: 'Inhaltsblock',
+    addressItem: 'Adresse',
+    openingHourItem: 'Öffnungszeit',
+    locationItem: 'Ort',
+    accessibilityInformationItem: 'Barrierefreiheitsangabe',
+    priceInformationItem: 'Preiseintrag',
+  },
+  values: {
+    notAvailable: 'Nicht verfügbar',
+    mediaContentTypes: {
+      unspecified: 'Nicht festgelegt',
+      image: 'Bild',
+      audio: 'Audio',
+      video: 'Video',
+      logo: 'Logo',
+      attachment: 'Anhang',
+    },
+    weekdays: {
+      MO: 'Montag',
+      TU: 'Dienstag',
+      WE: 'Mittwoch',
+      TH: 'Donnerstag',
+      FR: 'Freitag',
+      SA: 'Samstag',
+      SU: 'Sonntag',
+    },
+  },
+  richText: {
+    blockType: 'Textformat',
+    paragraph: 'Absatz',
+    heading2: 'Überschrift 2',
+    heading3: 'Überschrift 3',
+    blockquote: 'Zitat',
+    bulletList: 'Aufzählung',
+    orderedList: 'Nummerierung',
+    bold: 'Fett',
+    italic: 'Kursiv',
+    undo: 'Zurück',
+    redo: 'Vorwärts',
+    linkInput: 'Link-URL',
+    applyLink: 'Link setzen',
+  },
+  empty: {
+    title: 'Noch keine Generic Items vorhanden',
+  },
+  pagination: {
+    ariaLabel: 'Generic-Items-Pagination',
+    previous: 'Zurück',
+    next: 'Weiter',
+    pageLabel: 'Seite {{page}}',
+  },
+  messages: {
+    loading: 'Generic Items werden geladen.',
+    loadError: 'Generic Items konnten nicht geladen werden.',
+    missingContent: 'Das angeforderte Generic Item konnte nicht geladen werden.',
+    saveError: 'Generic Item konnte nicht gespeichert werden.',
+    createSuccess: 'Generic Item wurde erstellt.',
+    updateSuccess: 'Generic Item wurde aktualisiert.',
+    imagePickerEmpty: 'Keine passenden Bilder gefunden.',
+    mediaUploadInitializing: 'Upload wird vorbereitet.',
+    mediaUploadUploading: 'Bild wird hochgeladen.',
+    mediaUploadFinalizing: 'Bild wird eingebunden.',
+    mediaUploadSuccess: 'Bild wurde hinzugefügt.',
+    mediaUploadError: 'Bild konnte nicht hochgeladen werden.',
+    mediaUploadUnsupportedType: 'Dieser Dateityp wird nicht unterstützt.',
+    mediaUploadUnavailableUrl: 'Dieses Medium hat keine öffentliche URL.',
+    categoryOptionsLoading: 'Kategorien werden geladen.',
+    categoryOptionsLoadError: 'Kategorien konnten nicht geladen werden.',
+    locationGeocodeDisabled: 'Geo-Koordinaten sind für diese Instanz derzeit nicht verfügbar.',
+    locationGeocodeEmpty: 'Keine Koordinaten gefunden.',
+    locationGeocodeRateLimited: 'Geocoding-Limit erreicht.',
+    locationGeocodeTimeout: 'Geocoding hat zu lange gedauert.',
+    locationGeocodeForbidden: 'Berechtigung für Geocoding fehlt.',
+    locationGeocodeUnauthorized: 'Geocoding-Sitzung abgelaufen.',
+    locationGeocodeError: 'Geo-Koordinaten konnten nicht ermittelt werden.',
+    locationMapUnavailable: 'Karte nicht verfügbar.',
+    locationMapError: 'Karte konnte nicht geladen werden.',
+  },
+  validation: {
+    geoLocation: 'Koordinaten sind ungültig.',
+    categories: 'Kategorien benötigen einen Namen mit maximal 128 Zeichen.',
+    priceInformations: 'Preisangaben müssen valide Zahlen enthalten.',
+    webUrls: 'URLs müssen mit https:// beginnen.',
+  },
+} as const;
