+++ packages/plugin-generic-items/src/plugin.translations.de.ts
@@ -1,0 +1,171 @@
+import { genericItemsTranslationsDeActions } from './plugin.translations.de.actions.js';
+import { genericItemsTranslationsDeMessages } from './plugin.translations.de.messages.js';
+
+export const genericItemsTranslationsDe = {
+  genericItems: {
+    navigation: {
+      title: 'Generic Items',
+    },
+    list: {
+      title: 'Generic Items',
+      description: 'Freie Generic-Item-Inhalte im Studio verwalten.',
+    },
+    editor: {
+      createTitle: 'Generic Item anlegen',
+      createDescription: 'Erstellen Sie einen freien Generic-Item-Eintrag.',
+      editTitle: 'Generic Item bearbeiten',
+      editDescription: 'Aktualisieren oder löschen Sie den Generic-Item-Eintrag.',
+    },
+    fields: {
+      actions: 'Aktionen',
+      title: 'Titel',
+      genericType: 'Generic-Type',
+      visible: 'Sichtbar',
+      author: 'Autor',
+      keywords: 'Schlagwörter',
+      externalId: 'Externe ID',
+      publicationDate: 'Publikationsdatum',
+      publishedAt: 'Veröffentlicht am',
+      teaser: 'Teaser',
+      intro: 'Intro',
+      body: 'Inhalt',
+      categoryName: 'Primärkategorie',
+      categories: 'Kategorien',
+      categoriesHelp: 'Wählen Sie keine, eine oder mehrere Kategorien aus.',
+      categoriesSearch: 'Kategorien suchen',
+      categoriesSearchPlaceholder: 'Kategorie suchen oder auswählen',
+      contacts: 'Kontakte',
+      webUrls: 'Web-Links',
+      addresses: 'Adressen',
+      addressAddition: 'Adresszusatz',
+      addressKind: 'Adressart',
+      street: 'Straße',
+      zip: 'PLZ',
+      city: 'Ort',
+      latitude: 'Breitengrad',
+      longitude: 'Längengrad',
+      contentBlocks: 'Content-Blocks',
+      openingHours: 'Öffnungszeiten',
+      dateFrom: 'Datum von',
+      dateTo: 'Datum bis',
+      timeFrom: 'Uhrzeit von',
+      timeTo: 'Uhrzeit bis',
+      description: 'Beschreibung',
+      open: 'Geöffnet',
+      mediaContents: 'Medien',
+      mediaCaption: 'Bildunterschrift',
+      mediaCopyright: 'Copyright',
+      mediaContentType: 'Medientyp',
+      imageSearch: 'Bildsuche',
+      locations: 'Orte',
+      locationName: 'Ortsname',
+      department: 'Bereich',
+      district: 'Bezirk',
+      regionName: 'Region',
+      state: 'Bundesland/Land',
+      dates: 'Termine',
+      dateStart: 'Start',
+      dateEnd: 'Ende',
+      timeStart: 'Beginn',
+      timeEnd: 'Ende Uhrzeit',
+      weekday: 'Wochentag',
+      timeDescription: 'Zeitbeschreibung',
+      useOnlyTimeDescription: 'Nur Zeitbeschreibung verwenden',
+      accessibilityInformations: 'Barrierefreiheit',
+      accessibilityTypes: 'Barrierefreiheits-Typen',
+      accessibilityLinks: 'Weiterführende Links',
+      priceInformations: 'Preise',
+      priceName: 'Preisname',
+      priceAmount: 'Betrag',
+      priceCategory: 'Preiskategorie',
+      priceDescription: 'Preisbeschreibung',
+      groupPrice: 'Gruppenpreis',
+      ageFrom: 'Alter von',
+      ageTo: 'Alter bis',
+      minAdultCount: 'Erwachsene mindestens',
+      maxAdultCount: 'Erwachsene höchstens',
+      minChildrenCount: 'Kinder mindestens',
+      maxChildrenCount: 'Kinder höchstens',
+      payload: 'Payload',
+      updatedAt: 'Geändert am',
+      firstName: 'Vorname',
+      lastName: 'Nachname',
+      email: 'E-Mail',
+      phone: 'Telefon',
+      url: 'URL',
+      urlDescription: 'Linkbeschreibung',
+    },
+    tabs: {
+      ariaLabel: 'Detailbereiche',
+      basis: {
+        label: 'Basis',
+        title: 'Basis',
+        description: 'Identität, Typ und Kernmetadaten des Generic Items.',
+      },
+      content: {
+        label: 'Inhalt',
+        title: 'Inhalt',
+        description: 'Redaktionelle Inhalte und fachliche Feldgruppen des Generic Items.',
+      },
+      settings: {
+        label: 'Einstellungen',
+        title: 'Einstellungen',
+        description: 'Freie Payload und ergänzende technische Felder.',
+      },
+      history: {
+        label: 'Historie',
+        title: 'Historie',
+        description: 'Historienansicht für spätere Ausbaustufen.',
+      },
+    },
+    cards: {
+      basis: {
+        identity: {
+          title: 'Identität',
+          description: 'Titel, Typ und Sichtbarkeit des Eintrags.',
+        },
+        meta: {
+          title: 'Metadaten',
+          description: 'Autorschaft und Veröffentlichungsmerkmale.',
+        },
+      },
+      content: {
+        text: {
+          title: 'Text & Blöcke',
+          description: 'Kurzeinführung und strukturierte Inhaltsblöcke.',
+        },
+        classification: {
+          title: 'Klassifikation',
+          description: 'Kategorien und redaktionelle Zuordnung.',
+        },
+        relations: {
+          title: 'Kontakte & Orte',
+          description: 'Kontakte, Adressen und Ortsbezüge in einem Bereich.',
+        },
+        linksMedia: {
+          title: 'Links & Medien',
+          description: 'Externe Verweise und Medienobjekte.',
+        },
+        schedule: {
+          title: 'Termine & Öffnungszeiten',
+          description: 'Terminreihen und wiederkehrende Öffnungszeiten.',
+        },
+      },
+      settings: {
+        payload: {
+          title: 'Payload',
+          description: 'Freies JSON für plugin-spezifische Zusatzdaten.',
+        },
+        secondary: {
+          title: 'Sekundäre Felder',
+          description: 'Barrierefreiheits- und Preisinformationen im ersten Schnitt.',
+        },
+      },
+    },
+    history: {
+      placeholder: 'Die Historie wird in einem späteren Schritt an eine Host-Ansicht angeschlossen.',
+    },
+    ...genericItemsTranslationsDeActions,
+    ...genericItemsTranslationsDeMessages,
+  },
+} as const;
