+++ packages/plugin-generic-items/src/plugin.translations.en.ts
@@ -1,0 +1,5 @@
+export const genericItemsTranslationsEn = {
+  genericItems: {
+    navigation: { title: 'Generic Items' },
+  },
+} as const;
