+++ packages/plugin-generic-items/src/plugin.translations.ts
@@ -1,293 +1,6 @@
 import type { PluginTranslations } from '@sva/plugin-sdk';
-
-const genericItemsTranslationsDe = {
-  genericItems: {
-    navigation: {
-      title: 'Generic Items',
-    },
-    list: {
-      title: 'Generic Items',
-      description: 'Freie Generic-Item-Inhalte im Studio verwalten.',
-    },
-    editor: {
-      createTitle: 'Generic Item anlegen',
-      createDescription: 'Erstellen Sie einen freien Generic-Item-Eintrag.',
-      editTitle: 'Generic Item bearbeiten',
-      editDescription: 'Aktualisieren oder löschen Sie den Generic-Item-Eintrag.',
-    },
-    fields: {
-      actions: 'Aktionen',
-      title: 'Titel',
-      genericType: 'Generic-Type',
-      visible: 'Sichtbar',
-      author: 'Autor',
-      keywords: 'Schlagwörter',
-      externalId: 'Externe ID',
-      publicationDate: 'Publikationsdatum',
-      publishedAt: 'Veröffentlicht am',
-      teaser: 'Teaser',
-      intro: 'Intro',
-      body: 'Inhalt',
-      categoryName: 'Primärkategorie',
-      categories: 'Kategorien',
-      categoriesHelp: 'Wählen Sie keine, eine oder mehrere Kategorien aus.',
-      categoriesSearch: 'Kategorien suchen',
-      categoriesSearchPlaceholder: 'Kategorie suchen oder auswählen',
-      contacts: 'Kontakte',
-      webUrls: 'Web-Links',
-      addresses: 'Adressen',
-      addressAddition: 'Adresszusatz',
-      addressKind: 'Adressart',
-      street: 'Straße',
-      zip: 'PLZ',
-      city: 'Ort',
-      latitude: 'Breitengrad',
-      longitude: 'Längengrad',
-      contentBlocks: 'Content-Blocks',
-      openingHours: 'Öffnungszeiten',
-      dateFrom: 'Datum von',
-      dateTo: 'Datum bis',
-      timeFrom: 'Uhrzeit von',
-      timeTo: 'Uhrzeit bis',
-      description: 'Beschreibung',
-      open: 'Geöffnet',
-      mediaContents: 'Medien',
-      mediaCaption: 'Bildunterschrift',
-      mediaCopyright: 'Copyright',
-      mediaContentType: 'Medientyp',
-      imageSearch: 'Bildsuche',
-      locations: 'Orte',
-      locationName: 'Ortsname',
-      department: 'Bereich',
-      district: 'Bezirk',
-      regionName: 'Region',
-      state: 'Bundesland/Land',
-      dates: 'Termine',
-      dateStart: 'Start',
-      dateEnd: 'Ende',
-      timeStart: 'Beginn',
-      timeEnd: 'Ende Uhrzeit',
-      weekday: 'Wochentag',
-      timeDescription: 'Zeitbeschreibung',
-      useOnlyTimeDescription: 'Nur Zeitbeschreibung verwenden',
-      accessibilityInformations: 'Barrierefreiheit',
-      accessibilityTypes: 'Barrierefreiheits-Typen',
-      accessibilityLinks: 'Weiterführende Links',
-      priceInformations: 'Preise',
-      priceName: 'Preisname',
-      priceAmount: 'Betrag',
-      priceCategory: 'Preiskategorie',
-      priceDescription: 'Preisbeschreibung',
-      groupPrice: 'Gruppenpreis',
-      ageFrom: 'Alter von',
-      ageTo: 'Alter bis',
-      minAdultCount: 'Erwachsene mindestens',
-      maxAdultCount: 'Erwachsene höchstens',
-      minChildrenCount: 'Kinder mindestens',
-      maxChildrenCount: 'Kinder höchstens',
-      payload: 'Payload',
-      updatedAt: 'Geändert am',
-      firstName: 'Vorname',
-      lastName: 'Nachname',
-      email: 'E-Mail',
-      phone: 'Telefon',
-      url: 'URL',
-      urlDescription: 'Linkbeschreibung',
-    },
-    tabs: {
-      ariaLabel: 'Detailbereiche',
-      basis: {
-        label: 'Basis',
-        title: 'Basis',
-        description: 'Identität, Typ und Kernmetadaten des Generic Items.',
-      },
-      content: {
-        label: 'Inhalt',
-        title: 'Inhalt',
-        description: 'Redaktionelle Inhalte und fachliche Feldgruppen des Generic Items.',
-      },
-      settings: {
-        label: 'Einstellungen',
-        title: 'Einstellungen',
-        description: 'Freie Payload und ergänzende technische Felder.',
-      },
-      history: {
-        label: 'Historie',
-        title: 'Historie',
-        description: 'Historienansicht für spätere Ausbaustufen.',
-      },
-    },
-    cards: {
-      basis: {
-        identity: {
-          title: 'Identität',
-          description: 'Titel, Typ und Sichtbarkeit des Eintrags.',
-        },
-        meta: {
-          title: 'Metadaten',
-          description: 'Autorschaft und Veröffentlichungsmerkmale.',
-        },
-      },
-      content: {
-        text: {
-          title: 'Text & Blöcke',
-          description: 'Kurzeinführung und strukturierte Inhaltsblöcke.',
-        },
-        classification: {
-          title: 'Klassifikation',
-          description: 'Kategorien und redaktionelle Zuordnung.',
-        },
-        relations: {
-          title: 'Kontakte & Orte',
-          description: 'Kontakte, Adressen und Ortsbezüge in einem Bereich.',
-        },
-        linksMedia: {
-          title: 'Links & Medien',
-          description: 'Externe Verweise und Medienobjekte.',
-        },
-        schedule: {
-          title: 'Termine & Öffnungszeiten',
-          description: 'Terminreihen und wiederkehrende Öffnungszeiten.',
-        },
-      },
-      settings: {
-        payload: {
-          title: 'Payload',
-          description: 'Freies JSON für plugin-spezifische Zusatzdaten.',
-        },
-        secondary: {
-          title: 'Sekundäre Felder',
-          description: 'Barrierefreiheits- und Preisinformationen im ersten Schnitt.',
-        },
-      },
-    },
-    history: {
-      placeholder: 'Die Historie wird in einem späteren Schritt an eine Host-Ansicht angeschlossen.',
-    },
-    actions: {
-      create: 'Generic Item anlegen',
-      update: 'Änderungen speichern',
-      save: 'Speichern',
-      back: 'Zurück zur Liste',
-      edit: 'Bearbeiten',
-      delete: 'Löschen',
-      deleteConfirm: 'Soll dieses Generic Item wirklich gelöscht werden?',
-      addContentBlock: 'Block hinzufügen',
-      addCategory: 'Kategorie hinzufügen',
-      addAddress: 'Adresse hinzufügen',
-      addOpeningHour: 'Öffnungszeit hinzufügen',
-      addLink: 'Link hinzufügen',
-      addContact: 'Kontakt hinzufügen',
-      addDate: 'Termin hinzufügen',
-      geocodeAddress: 'Geo-Koordinaten ermitteln',
-      geocodingAddress: 'Geo-Koordinaten werden ermittelt',
-      reverseGeocodeAddress: 'Adresse ermitteln',
-      reverseGeocodingAddress: 'Adresse wird ermittelt',
-      addImage: 'Aus Medienbibliothek wählen',
-      addLocation: 'Ort hinzufügen',
-      addAccessibilityInformation: 'Barrierefreiheit hinzufügen',
-      addPriceInformation: 'Preis hinzufügen',
-      uploadMedia: 'Bild hochladen',
-      uploadingMedia: 'Bild wird hochgeladen',
-      addMediaManual: 'Medium manuell ergänzen',
-      selectImage: 'Bild auswählen',
-      removeImage: 'Bild entfernen',
-      removeCategory: 'Kategorie {{name}} entfernen',
-      remove: 'Entfernen',
-    },
-    content: {
-      blockItem: 'Inhaltsblock',
-      addressItem: 'Adresse',
-      openingHourItem: 'Öffnungszeit',
-      locationItem: 'Ort',
-      accessibilityInformationItem: 'Barrierefreiheitsangabe',
-      priceInformationItem: 'Preiseintrag',
-    },
-    values: {
-      notAvailable: 'Nicht verfügbar',
-      mediaContentTypes: {
-        unspecified: 'Nicht festgelegt',
-        image: 'Bild',
-        audio: 'Audio',
-        video: 'Video',
-        logo: 'Logo',
-        attachment: 'Anhang',
-      },
-      weekdays: {
-        MO: 'Montag',
-        TU: 'Dienstag',
-        WE: 'Mittwoch',
-        TH: 'Donnerstag',
-        FR: 'Freitag',
-        SA: 'Samstag',
-        SU: 'Sonntag',
-      },
-    },
-    richText: {
-      blockType: 'Textformat',
-      paragraph: 'Absatz',
-      heading2: 'Überschrift 2',
-      heading3: 'Überschrift 3',
-      blockquote: 'Zitat',
-      bulletList: 'Aufzählung',
-      orderedList: 'Nummerierung',
-      bold: 'Fett',
-      italic: 'Kursiv',
-      undo: 'Zurück',
-      redo: 'Vorwärts',
-      linkInput: 'Link-URL',
-      applyLink: 'Link setzen',
-    },
-    empty: {
-      title: 'Noch keine Generic Items vorhanden',
-    },
-    pagination: {
-      ariaLabel: 'Generic-Items-Pagination',
-      previous: 'Zurück',
-      next: 'Weiter',
-      pageLabel: 'Seite {{page}}',
-    },
-    messages: {
-      loading: 'Generic Items werden geladen.',
-      loadError: 'Generic Items konnten nicht geladen werden.',
-      missingContent: 'Das angeforderte Generic Item konnte nicht geladen werden.',
-      saveError: 'Generic Item konnte nicht gespeichert werden.',
-      createSuccess: 'Generic Item wurde erstellt.',
-      updateSuccess: 'Generic Item wurde aktualisiert.',
-      imagePickerEmpty: 'Keine passenden Bilder gefunden.',
-      mediaUploadInitializing: 'Upload wird vorbereitet.',
-      mediaUploadUploading: 'Bild wird hochgeladen.',
-      mediaUploadFinalizing: 'Bild wird eingebunden.',
-      mediaUploadSuccess: 'Bild wurde hinzugefügt.',
-      mediaUploadError: 'Bild konnte nicht hochgeladen werden.',
-      mediaUploadUnsupportedType: 'Dieser Dateityp wird nicht unterstützt.',
-      mediaUploadUnavailableUrl: 'Dieses Medium hat keine öffentliche URL.',
-      categoryOptionsLoading: 'Kategorien werden geladen.',
-      categoryOptionsLoadError: 'Kategorien konnten nicht geladen werden.',
-      locationGeocodeDisabled: 'Geo-Koordinaten sind für diese Instanz derzeit nicht verfügbar.',
-      locationGeocodeEmpty: 'Keine Koordinaten gefunden.',
-      locationGeocodeRateLimited: 'Geocoding-Limit erreicht.',
-      locationGeocodeTimeout: 'Geocoding hat zu lange gedauert.',
-      locationGeocodeForbidden: 'Berechtigung für Geocoding fehlt.',
-      locationGeocodeUnauthorized: 'Geocoding-Sitzung abgelaufen.',
-      locationGeocodeError: 'Geo-Koordinaten konnten nicht ermittelt werden.',
-      locationMapUnavailable: 'Karte nicht verfügbar.',
-      locationMapError: 'Karte konnte nicht geladen werden.',
-    },
-    validation: {
-      geoLocation: 'Koordinaten sind ungültig.',
-      categories: 'Kategorien benötigen einen Namen mit maximal 128 Zeichen.',
-      priceInformations: 'Preisangaben müssen valide Zahlen enthalten.',
-      webUrls: 'URLs müssen mit https:// beginnen.',
-    },
-  },
-} as const;
-
-const genericItemsTranslationsEn = {
-  genericItems: {
-    navigation: { title: 'Generic Items' },
-  },
-} as const;
+import { genericItemsTranslationsDe } from './plugin.translations.de.js';
+import { genericItemsTranslationsEn } from './plugin.translations.en.js';
 
 export const pluginGenericItemsTranslations = {
   de: genericItemsTranslationsDe,
