+++ packages/sva-mainserver/src/server/generic-items-route-input.dates.ts
@@ -1,0 +1,20 @@
+import type { SvaMainserverDateInput } from '../types.js';
+import { isRecord, readBoolean, readString } from './content-route-core.js';
+
+export const parseDates = (value: unknown): readonly SvaMainserverDateInput[] | undefined =>
+  Array.isArray(value)
+    ? value
+        .filter(isRecord)
+        .map((date): SvaMainserverDateInput => ({
+          ...(readString(date.weekday) ? { weekday: readString(date.weekday) } : {}),
+          ...(readString(date.dateStart) ? { dateStart: readString(date.dateStart) } : {}),
+          ...(readString(date.dateEnd) ? { dateEnd: readString(date.dateEnd) } : {}),
+          ...(readString(date.timeStart) ? { timeStart: readString(date.timeStart) } : {}),
+          ...(readString(date.timeEnd) ? { timeEnd: readString(date.timeEnd) } : {}),
+          ...(readString(date.timeDescription) ? { timeDescription: readString(date.timeDescription) } : {}),
+          ...(readBoolean(date.useOnlyTimeDescription) !== undefined
+            ? { useOnlyTimeDescription: readBoolean(date.useOnlyTimeDescription) }
+            : {}),
+        }))
+        .filter((date) => Object.keys(date).length > 0)
+    : undefined;
